feat(air-parser)!: Refactor Instruction type
Instead of being boxed for each node, the `Instruction` type contains a
boxed variable-size elements.  Thus `Instruction` is quite lean, and the
allocator deals with variable-sized elements.

Total number of allocations is more or less same, but less space is
wasted for unused memory: previously the Instruction's size was 112
bytes in WASM, now it is 16.

It reduces memory consumption on large AIR scripts (heap size decreased
from 7.7Mb to 4.9Mb in parser-10000-100, for example).